--- app/models/Request.scala
@@ -1,7 +1,0 @@
-package models
-
-case class Request(
-  sequence: Int,
-  sql: String
-)
-
