+++ app/models/ResponseTable.scala
@@ -4,11 +4,13 @@ import play.api.libs.json.JsString
 import play.api.libs.json.Json
 import play.api.libs.json.OWrites
 
+import scala.collection.immutable
+
 case class ResponseTable(
   sequence: Int,
   sql: String,
   mode: Int = 1,
-  data: Seq[JsString]
+  data: immutable.Seq[String]
 )
 
 object ResponseTable {
